requirements.txt
streamlit==1.29.0
pillow
pandas
streamlit-image-coordinates==0.1.5